Complex Deletion

+++ src/editor/view.rs
@@ -34,7 +34,6 @@ impl View {
             EditorCommand::Delete => self.delete(),
             EditorCommand::Backspace => self.backspace(),
         }
-        self.scroll_text_location_into_view();
     }
 
     pub fn load(&mut self, file_name: &str) {
@@ -46,20 +45,24 @@ impl View {
 
     fn resize(&mut self, to: Size) {
         self.needs_redraw = true;
+        self.scroll_text_location_into_view();
         self.size = to;
     }
 
     // region: Text-editing
 
     fn backspace(&mut self) {
-        self.move_left();
-        self.delete();
+        if self.text_location.line_index != 0 || self.text_location.grapheme_index != 0 {
+            self.move_text_location(&Direction::Left);
+            self.delete();
+        }
     }
 
     fn delete(&mut self) {
         self.buffer.delete_char(self.text_location);
         self.needs_redraw = true;
     }
+
     fn insert_char(&mut self, character: char) {
         let old_len = self
             .buffer
@@ -75,7 +78,7 @@ impl View {
         let grapheme_delta = new_len.saturating_sub(old_len);
         if grapheme_delta > 0 {
             // move right for an added grapheme (should be the regular case)
-            self.move_right();
+            self.move_text_location(&Direction::Right);
         }
         self.needs_redraw = true;
     }
@@ -213,6 +216,7 @@ impl View {
             Direction::Home => self.move_to_start_of_line(),
             Direction::End => self.move_to_end_of_line(),
         }
+        self.scroll_text_location_into_view();
     }
 
     fn move_up(&mut self, step: usize) {
